+++ csim/language/parser.py
@@ -7,6 +7,8 @@ from ..java_24.Java24Parser import Java24Parser
 from ..java_24.Java24Lexer import Java24Lexer
 from ..cpp_14.CPP14Parser import CPP14Parser
 from ..cpp_14.CPP14Lexer import CPP14Lexer
+from ..python_3.Python3Parser import Python3Parser
+from ..python_3.Python3Lexer import Python3Lexer
 from antlr4 import InputStream, CommonTokenStream
 from antlr4.error.ErrorListener import ErrorListener
 
@@ -92,6 +94,17 @@ def ANTLR_parse(file_name, file_content, lang):
         parser.removeErrorListeners()
         parser.addErrorListener(error_listener)
         tree = parser.translationUnit()
+    elif lang == "python_3":
+        # Lexing the input code to create a token stream
+        lexer = Python3Lexer(input_stream)
+        lexer.removeErrorListeners()
+        lexer.addErrorListener(error_listener)
+        # Parsing the token stream to create a parse tree
+        token_stream = CommonTokenStream(lexer)
+        parser = Python3Parser(token_stream)
+        parser.removeErrorListeners()
+        parser.addErrorListener(error_listener)
+        tree = parser.file_input()
     else:
         raise ValueError(f"Unsupported language: {lang}")